+++ CHANGELOG.md
@@ -4,15 +4,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.0.0
+
+### Changed
+
+- package name to `swift-version-compare` to be in line with community standards
+
+### Removed
+
+- [CocoaPods](https://cocoapods.org) support
+
 ## [1.2.0] - 2024-02-03
 
 ### Changed
 
-- `swift-tools-version` to `5.9.2` for the package manifest and tools
+- `swift-tools-version` to `5.7` for the package manifest and tools
 
 ### Removed
 
 - generated .xcodeproj file
+- codecov
 
 ## [1.1.0] - 2021-10-02